requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
Cython==0.29.34
docutils==0.19
exceptiongroup==1.1.1
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.2
h11==0.14.0
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.4
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
kivymd==1.1.1
lxml==4.9.2
magic-filter==1.0.9
MarkupSafe==2.1.2
multidict==6.0.4
mysql-connector-python==8.0.33
numpy==1.24.2
outcome==1.2.0
pandas==1.5.3
Pillow==9.5.0
protobuf==3.20.3
psycopg2==2.9.5
pycparser==2.21
Pygments==2.14.0
PyMySQL==1.0.2
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==306
requests==2.28.2
selectolax==0.3.12
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
Tcl==0.2
tk==0.1.0
transliterate==1.10.2
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
Werkzeug==2.2.3
wsproto==1.2.0
yarl==1.8.2
zope.event==4.6
zope.interface==6.0
